# backend/openapi-golden.json
{
  "components": {
    "schemas": {
      "Game": {
        "properties": {
          "away_team_abbrev": {
            "title": "Away Team Abbrev",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "game_date": {
            "format": "date",
            "title": "Game Date",
            "type": "string"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          },
          "home_team_abbrev": {
            "title": "Home Team Abbrev",
            "type": "string"
          },
          "season_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Year"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "game_date",
          "home_team_abbrev",
          "away_team_abbrev"
        ],
        "title": "Game",
        "type": "object"
      },
...
      "GameSlate": {
        "properties": {
          "game_date": {
            "format": "date",
            "title": "Game Date",
            "type": "string"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/Game"
            },
            "title": "Games",
            "type": "array"
          },
          "props": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PropLine"
            },
            "title": "Props",
            "type": "array"
          }
        },
        "required": [
          "game_date",
          "games"
        ],
        "title": "GameSlate",
        "type": "object"
      },
      "GameWithProps": {
        "properties": {
          "away_team_abbrev": {
            "title": "Away Team Abbrev",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "game_date": {
            "format": "date",
            "title": "Game Date",
            "type": "string"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          },
          "home_team_abbrev": {
            "title": "Home Team Abbrev",
            "type": "string"
          },
          "props": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PropLine"
            },
            "title": "Props",
            "type": "array"
          },
          "season_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Year"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "game_date",
          "home_team_abbrev",
          "away_team_abbrev"
        ],
        "title": "GameWithProps",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlayerListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            },
            "title": "Players",
            "type": "array"
          }
        },
        "required": [
          "count",
          "players"
        ],
        "title": "PlayerListResponse",
        "type": "object"
      },
      "PlayerProfile": {
        "description": "Full player response \u2014 profile + recent games + rolling context.",
        "properties": {
          "career_game_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Career Game Count"
          },
          "normalized_name": {
            "title": "Normalized Name",
            "type": "string"
          },
          "player_id": {
            "title": "Player Id",
            "type": "integer"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "recent_games": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlayerGame"
            },
            "title": "Recent Games",
            "type": "array"
          },
          "rolling_avg_10": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RollingAvg10"
              },
              {
                "type": "null"
              }
            ]
          },
          "rolling_avg_5": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RollingAvg5"
              },
              {
                "type": "null"
              }
            ]
          },
          "team_abbreviation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Abbreviation"
          },
          "team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          }
        },
        "required": [
          "player_id",
          "player_name",
          "normalized_name"
        ],
        "title": "PlayerProfile",
        "type": "object"
      },
...
      "PropLine": {
        "description": "One prop quote from silver.silver_props or gold.gold_prop_history.",
        "properties": {
          "ast_per_min_roll5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ast Per Min Roll5"
          },
          "away_team_abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Team Abbrev"
          },
          "bookmaker": {
            "title": "Bookmaker",
            "type": "string"
          },
          "expected_pace": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Pace"
          },
          "game_date": {
            "format": "date",
            "title": "Game Date",
            "type": "string"
          },
          "game_season_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Season Year"
          },
          "game_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Total"
          },
          "home_team_abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Team Abbrev"
          },
          "last_update_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Update At"
          },
          "line": {
            "title": "Line",
            "type": "number"
          },
          "market_category": {
            "title": "Market Category",
            "type": "string"
          },
          "min_roll10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Roll10"
          },
          "min_roll5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Roll5"
          },
          "normalized_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Name"
          },
          "odds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odds"
          },
          "opp_def_rating_roll10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opp Def Rating Roll10"
          },
          "player_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Id"
          },
          "player_name": {
            "title": "Player Name",
            "type": "string"
          },
          "player_name_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Name Raw"
          },
          "player_team_abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Team Abbrev"
          },
          "prop_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prop Source"
          },
          "pts_per_min_roll10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pts Per Min Roll10"
          },
          "pts_per_min_roll5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pts Per Min Roll5"
          },
          "reb_per_min_roll5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reb Per Min Roll5"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "team_min_rank_l10": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Min Rank L10"
          },
          "team_spread": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Spread"
          },
          "team_usg_rank_l10": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Usg Rank L10"
          }
        },
        "required": [
          "bookmaker",
          "market_category",
          "player_name",
          "side",
          "game_date",
          "line"
        ],
        "title": "PropLine",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Statvista basketball and baseball data backend. Database-backed endpoints read from Supabase (silver / gold schemas) and make no NBA or Odds API calls. The exceptions are /api/wnba/scoreboard/today, /api/mlb/scoreboard/today, /api/mlb/scoreboard, /api/mlb/odds/today, /api/wnba/leaders, /api/wnba/player/{player_id}, /api/wnba/standings, /api/wnba/futures, /api/wnba/odds/today, /api/wnba/props/today, WNBA game detail routes, /api/mlb/games/{game_pk}, and /api/mlb/lineups, which call ESPN, stats.wnba.com, MLB Stats API, RotoWire, or SharpAPI for live league data.",
    "title": "statvista API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/games/{date}": {
      "get": {
        "description": "Return all games on *date* (YYYY-MM-DD) from **silver.silver_games**.",
        "operationId": "get_games_api_games__date__get",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  },
                  "title": "Response Get Games Api Games  Date  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Games",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{date}/props": {
      "get": {
        "description": "All prop lines for a slate date from **gold.gold_prop_history**.",
        "operationId": "get_game_props_api_games__date__props_get",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "bookmaker",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bookmaker"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropLine"
                  },
                  "title": "Response Get Game Props Api Games  Date  Props Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Game Props",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{date}/slate": {
      "get": {
        "description": "Combined games + props for a full slate view.",
        "operationId": "get_game_slate_api_games__date__slate_get",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSlate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Game Slate",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{date}/with-props": {
      "get": {
        "description": "Games on *date* with prop lines grouped per matchup.",
        "operationId": "get_games_with_props_api_games__date__with_props_get",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "bookmaker",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bookmaker"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GameWithProps"
                  },
                  "title": "Response Get Games With Props Api Games  Date  With Props Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Games With Props",
        "tags": [
          "games"
        ]
      }
    },
...
    "/api/mlb/games/{game_pk}": {
      "get": {
        "operationId": "mlb_game_detail_api_mlb_games__game_pk__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_pk",
            "required": true,
            "schema": {
              "title": "Game Pk",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MlbGameDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mlb Game Detail",
        "tags": [
          "mlb"
        ]
      }
    },
...
    "/api/player/{player_id}": {
      "get": {
        "description": "Player profile, recent games, and rolling context.",
        "operationId": "get_player_api_player__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "recent_n",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 82,
              "minimum": 1,
              "title": "Recent N",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player",
        "tags": [
          "players"
        ]
      }
    },
    "/api/players": {
      "get": {
        "description": "Search the player directory from **silver.silver_players**.",
        "operationId": "search_players_api_players_get",
        "parameters": [
          {
            "description": "Name search (case-insensitive).",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name search (case-insensitive).",
              "title": "Q"
            }
          },
          {
            "description": "Team tricode, e.g. LAL",
            "in": "query",
            "name": "team",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Team tricode, e.g. LAL",
              "title": "Team"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Players",
        "tags": [
          "players"
        ]
      }
    },
...
    "/api/wnba/games/{espn_event_id}": {
      "get": {
        "operationId": "wnba_game_detail_api_wnba_games__espn_event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "espn_event_id",
            "required": true,
            "schema": {
              "title": "Espn Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WnbaGameDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wnba Game Detail",
        "tags": [
          "wnba"
        ]
      }
    },
...
  }
}